Update lodash to latest version, refactor jasmine task.
So the newest version of lodash's merge method does deep cloning
of the sources it is merging, which causes some very weird issues when
deep cloning the jasmine browser plugin.

The workaround is just to have the configuration with the plugin not be
one of the object sources for lodash's _.merge.